Build scripts update

+++ build_5.7_debug.sh
@@ -23,13 +23,22 @@ cmake . -DWITH_ZLIB=system -DCMAKE_BUILD_TYPE=Debug -DBUILD_CONFIG=mysql_release
 make | tee -a /tmp/5.7_debug_build
 ./scripts/make_binary_distribution | tee -a /tmp/5.7_debug_build
 TAR_dbg=`ls -1 *.tar.gz | head -n1`
-TAR_dbg_new=$(echo "${1}-${TAR_dbg}" | sed 's|.tar.gz|-debug.tar.gz|')
-mv ${TAR_dbg} ../${TAR_dbg_new}
-cd ..
-tar -xf ${TAR_dbg_new}
-DIR_dbg=$(echo "${TAR_dbg}" | sed 's|.tar.gz||')
-DIR_dbg_new=$(echo "${TAR_dbg_new}" | sed 's|.tar.gz||')
-mv ${DIR_dbg} ${DIR_dbg_new}
-echo "Done! Now run;"
-echo "mv ../${DIR_dbg_new} /sda"  # The script will end still in $PWD, hence we will need ../ (output only)
-rm -Rf ${CURPATH}_dbg
+if [ "$TAR_dbg}" != "" ]; then
+  DIR_dbg=$(echo "${TAR_dbg}" | sed 's|.tar.gz||')
+  TAR_dbg_new=$(echo "${1}-${TAR_dbg}" | sed 's|.tar.gz|-debug.tar.gz|')
+  DIR_dbg_new=$(echo "${TAR_dbg_new}" | sed 's|.tar.gz||')
+  if [ "${DIR_dbg}" != "" ]; then rm -Rf ../${DIR_dbg}; fi
+  if [ "${DIR_dbg_new}" != "" ]; then rm -Rf ../${DIR_dbg_new}; fi
+  if [ "${TAR_dbg_new}" != "" ]; then rm -Rf ../${TAR_dbg_new}; fi
+  mv ${TAR_dbg} ../${TAR_dbg_new}
+  cd ..
+  tar -xf ${TAR_dbg_new}
+  mv ${DIR_dbg} ${DIR_dbg_new}
+  echo "Done! Now run;"
+  echo "mv ../${DIR_dbg_new} /sda"  # The script will end still in $PWD, hence we will need ../ (output only)
+  rm -Rf ${CURPATH}_dbg
+  exit 0
+else
+  echo "There was some build issue... Have a nice day!"
+  exit 1
+fi
